全面更新数据库，ID全部设置为bigint类型，并以id作为外键各表相连

+++ src/main/java/spring17/dto/CommentDTO.java
@@ -9,13 +9,13 @@ import spring17.model.User;
  */
 @Data
 public class CommentDTO {
-    private Integer id;
-    private Integer parentId;
+    private Long id;
+    private Long parentId;
     private Integer type;
-    private Integer commentator;
+    private Long commentator;
     private Long gmtCreate;
     private Long gmtModified;
-    private Long likeCount;
+    private Integer likeCount;
     private String content;
     //新增：
     private User user;
